Check document can update as in documents_items view; fixes #1171

+++ front/ticket.form.php
@@ -159,14 +159,17 @@ if (isset($_POST["add"])) {
               sprintf(__('%s adds an actor'), $_SESSION["glpiname"]));
    Html::redirect($CFG_GLPI["root_doc"]."/front/ticket.form.php?id=".$_POST['tickets_id']);
 } else if (isset($_REQUEST['delete_document'])) {
-   $document_item = new Document_Item;
-   $found_document_items = $document_item->find("itemtype = 'Ticket' ".
-                                                " AND items_id = ".intval($_REQUEST['tickets_id']).
-                                                " AND documents_id = ".intval($_REQUEST['documents_id']));
-   foreach ($found_document_items  as $item) {
-      $document_item->delete($item, true);
+   $doc = new Document();
+   $doc->getFromDB(intval($_REQUEST['documents_id']));
+   if ($doc->can($doc->getID(), UPDATE)) {
+      $document_item = new Document_Item;
+      $found_document_items = $document_item->find("itemtype = 'Ticket' ".
+                                                   " AND items_id = ".intval($_REQUEST['tickets_id']).
+                                                   " AND documents_id = ".$doc->getID());
+      foreach ($found_document_items  as $item) {
+         $document_item->delete($item, true);
+      }
    }
-
    Html::back();
 }
